requirements.txt
flask
chromadb
pdfplumber
# sudo pacman -S python-google-api-python-client python-google-auth-oauthlib
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
requests
